refactor(ledger): move transaction create core from http handler into command use case :recycle:
