+++ components/ledger/internal/adapters/http/in/observability_test.go
@@ -9,7 +9,6 @@ import (
 	"testing"
 	"time"
 
-	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
 	libHTTP "github.com/LerianStudio/midaz/v4/pkg/net/http"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -18,62 +17,6 @@ import (
 )
 
 // --- Onboarding observability tests ---
-
-func TestSafePayloadAttributes(t *testing.T) {
-	t.Parallel()
-
-	alias := "@ops"
-	portfolioID := "portfolio-id"
-	segmentID := "segment-id"
-	entityID := "entity-id"
-
-	tests := []struct {
-		name      string
-		payload   any
-		assertion func(t *testing.T, attrs map[string]any)
-	}{
-		{
-			name: "account payload only exposes safe presence flags",
-			payload: &mmodel.CreateAccountInput{
-				Alias:       &alias,
-				PortfolioID: &portfolioID,
-				SegmentID:   &segmentID,
-				EntityID:    &entityID,
-				Metadata:    map[string]any{"secret": "value"},
-			},
-			assertion: func(t *testing.T, attrs map[string]any) {
-				t.Helper()
-
-				assert.Equal(t, "CreateAccountInput", attrs["app.request.payload.type"])
-				assert.Equal(t, true, attrs["app.request.payload.has_metadata"])
-				assert.Equal(t, true, attrs["app.request.payload.has_portfolio_id"])
-				assert.Equal(t, true, attrs["app.request.payload.has_segment_id"])
-				assert.Equal(t, true, attrs["app.request.payload.has_entity_id"])
-				assert.Equal(t, true, attrs["app.request.payload.has_alias"])
-			},
-		},
-		{
-			name:    "nil payload is reported as unknown without extra data",
-			payload: nil,
-			assertion: func(t *testing.T, attrs map[string]any) {
-				t.Helper()
-
-				assert.Equal(t, "unknown", attrs["app.request.payload.type"])
-				assert.Equal(t, false, attrs["app.request.payload.has_metadata"])
-				assert.Equal(t, false, attrs["app.request.payload.has_alias"])
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			attrs := attributeMap(safePayloadAttributes(tt.payload))
-			tt.assertion(t, attrs)
-		})
-	}
-}
 
 func TestSafeQueryAttributes(t *testing.T) {
 	t.Parallel()
@@ -120,30 +63,6 @@ func TestSafeQueryAttributes_DefaultAndNilCases(t *testing.T) {
 	assert.Equal(t, "", defaultAttrs["app.request.query.sort_order"])
 	assert.Equal(t, false, defaultAttrs["app.request.query.has_cursor"])
 	assert.Equal(t, int64(0), defaultAttrs["app.request.query.to_asset_codes_count"])
-}
-
-func TestSafePayloadAttributes_RedactsValues(t *testing.T) {
-	t.Parallel()
-
-	payload := struct {
-		Alias          string
-		Key            string
-		OrganizationID string
-	}{
-		Alias:          "sensitive-alias",
-		Key:            "super-secret-key",
-		OrganizationID: "org-123",
-	}
-
-	attrs := safePayloadAttributes(payload)
-	require.NotEmpty(t, attrs)
-
-	for _, attr := range attrs {
-		serialized := fmt.Sprint(attr.Value.AsInterface())
-		assert.NotContains(t, serialized, "sensitive-alias")
-		assert.NotContains(t, serialized, "super-secret-key")
-		assert.NotContains(t, serialized, "org-123")
-	}
 }
 
 func TestSafeQueryAttributes_RedactsValues(t *testing.T) {
