+++ components/ledger/internal/adapters/http/in/transaction_core.go
@@ -22,25 +22,6 @@ import (
 type TransactionHandler struct {
 	Command *command.UseCase
 	Query   *query.UseCase
-	// FeeApplier drives the in-process fee engine inside the create seam. It is
-	// injected at bootstrap from the fee use case; a nil applier disables fee
-	// application (the create path stays unchanged).
-	FeeApplier FeeApplier
-	// TracerReserver drives the tracer two-phase reservation lifecycle from the
-	// create seam. It is injected at bootstrap from the tracer HTTP client; a
-	// nil reserver means the tracer integration is disabled (the create path
-	// stays unchanged). The per-ledger tracer.mode gate lives at the call site.
-	TracerReserver TracerReserver
-	// FeesMongoManager resolves the CURRENT tenant's fee Mongo database at the
-	// fee seam when MultiTenantEnabled is true. The fee pack/billing repos read
-	// the GENERIC tmcore MB key, which the route-scoped feesTenantMiddleware
-	// only sets on FEE routes — never on the transaction route — so the seam
-	// must resolve and inject it onto a derived ctx itself. Nil in single-tenant
-	// mode (and in tests that do not exercise the seam).
-	FeesMongoManager feesDBResolver
-	// MultiTenantEnabled gates the fee-seam tenant resolution. When false the
-	// static fee connection is correct and resolveFeesTenantContext is a no-op.
-	MultiTenantEnabled bool
 }
 
 // buildOverriddenTransaction builds the transaction from the input, forces
