+++ components/ledger/internal/adapters/http/in/transaction_primary_read_intent_test.go
@@ -5,393 +5,60 @@
 package in
 
 import (
-	"context"
 	"go/ast"
 	"go/parser"
 	"go/token"
 	"os"
 	"strings"
 	"testing"
-
-	libConstants "github.com/LerianStudio/lib-commons/v6/commons/constants"
-	"github.com/google/uuid"
-	"github.com/shopspring/decimal"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/mock/gomock"
-
-	"github.com/LerianStudio/midaz/v4/components/ledger/internal/adapters/postgres/balance"
-	redis "github.com/LerianStudio/midaz/v4/components/ledger/internal/adapters/redis/transaction"
-	"github.com/LerianStudio/midaz/v4/components/ledger/internal/services/query"
-	"github.com/LerianStudio/midaz/v4/components/ledger/pkg/readrouting"
-	"github.com/LerianStudio/midaz/v4/pkg/constant"
-	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
-	"github.com/LerianStudio/midaz/v4/pkg/mtransaction"
-	"github.com/LerianStudio/midaz/v4/pkg/utils"
 )
 
-// TestCreateTransaction_MarksPrimaryReadIntent verifies the create flow marks
-// the primary-read intent before its pre-write balance reads so the read-routing
-// seam can steer the DB miss to the primary. It asserts on the two facets that
-// make the mechanism real:
-//
-//  1. Mechanism: the intent marker survives the exact call sequence the handler
-//     uses. GetBalances (the direct pre-write read) and enrichOverdraftOperations
-//     (the second, overdraft-enrichment read that reuses the SAME ctx) both
-//     forward a marked ctx down to the seam target — the balance DB repository,
-//     which is where the read-routing seam observes the intent. A negative
-//     baseline proves an unmarked ctx stays unmarked, so the assertion is not
-//     trivially true.
-//  2. Placement: a structural guard over the live source of executeCreateTransaction
-//     proves the `readrouting.WithPrimaryRead(ctx)` wrap actually exists in the
-//     handler and positionally precedes the GetBalances read (and therefore the
-//     overdraft read, which follows it). Runtime mocks cannot catch a future
-//     removal or reorder of the wrap; the AST guard can. This mirrors the
-//     existing fee-seam structural guards (transaction_fee_seam_structure_test.go).
-func TestCreateTransaction_MarksPrimaryReadIntent(t *testing.T) {
-	t.Run("direct_balance_read_forwards_marked_ctx_to_db_seam", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
+// TestCommitCancel_PrimaryReadWrapPlacement is the placement half of the
+// commit/cancel primary-read contract: a structural guard over the live source of
+// commitOrCancelTransaction proving the dedicated-var wrap
+// `readCtx := readrouting.WithPrimaryRead(ctx)` exists AND sits AFTER the
+// validation-only reads (GetParsedLedgerSettings) and BEFORE the GetBalances read;
+// that GetBalances and the cancel overdraft read receive readCtx while
+// ValidateAccountingRules keeps the unmarked ctx, so only the pre-write balance reads
+// are routed to primary. The mechanism half lives beside the reads it exercises, in
+// the command package.
+func TestCommitCancel_PrimaryReadWrapPlacement(t *testing.T) {
+	src := readStateHandlerSource(t)
 
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
+	positions := analyzeCommitCancelWrap(t, src, "commitOrCancelTransaction")
 
-		organizationID := uuid.New()
-		ledgerID := uuid.New()
-		aliases := []string{"@alice#default"}
-
-		// Reproduce the handler's exact call sequence: mark the ctx, then read.
-		ctx := readrouting.WithPrimaryRead(context.Background())
-
-		_, err := uc.GetBalances(ctx, organizationID, ledgerID, aliases)
-		require.NoError(t, err)
-
-		require.True(t, captured.seen, "balance DB repository was never reached; the cache-miss path must fall through to the seam target")
-		assert.True(t, captured.primaryRead, "the ctx reaching the balance DB seam must carry the primary-read intent")
-	})
-
-	t.Run("overdraft_enrichment_read_forwards_marked_ctx_to_db_seam", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
-
-		handler := &TransactionHandler{Query: uc}
-
-		orgID := uuid.New()
-		ledgerID := uuid.New()
-
-		source := overdraftEnabledBalance(t, "@alice", decimal.NewFromInt(50), "100")
-
-		primary := mmodel.BalanceOperation{
-			Balance: source,
-			Alias:   "0#@alice#default",
-			Amount: mtransaction.Amount{
-				Asset:           "BRL",
-				Value:           decimal.NewFromInt(100),
-				Operation:       libConstants.DEBIT,
-				TransactionType: libConstants.CREATED,
-				Direction:       constant.DirectionCredit,
-			},
-			InternalKey: utils.BalanceInternalKey(orgID, ledgerID, "@alice#default"),
-		}
-
-		validate := &mtransaction.Responses{
-			From:    map[string]mtransaction.Amount{"0#@alice#default": primary.Amount},
-			Sources: []string{"@alice#default"},
-			Aliases: []string{"@alice#default"},
-		}
-
-		// enrichOverdraftOperations is invoked with the SAME wrapped ctx the
-		// handler holds, and the loader is handler.Query.GetBalances exactly as
-		// at the create call site. The overdraft read must therefore forward the
-		// mark down to the same DB seam target.
-		ctx := readrouting.WithPrimaryRead(context.Background())
-
-		_, _, err := enrichOverdraftOperations(ctx, orgID, ledgerID,
-			[]mmodel.BalanceOperation{primary}, validate, handler.Query.GetBalances)
-		require.NoError(t, err)
-
-		require.True(t, captured.seen, "overdraft companion read never reached the balance DB seam")
-		assert.True(t, captured.primaryRead, "the ctx reaching the balance DB seam via the overdraft read must carry the primary-read intent")
-	})
-
-	t.Run("unmarked_baseline_ctx_is_not_marked", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
-
-		_, err := uc.GetBalances(context.Background(), uuid.New(), uuid.New(), []string{"@alice#default"})
-		require.NoError(t, err)
-
-		require.True(t, captured.seen)
-		assert.False(t, captured.primaryRead, "a pure/unmarked read must NOT carry the primary-read intent")
-	})
-
-	t.Run("wrap_exists_and_precedes_get_balances_in_source", func(t *testing.T) {
-		src := readSeamSource(t)
-
-		wrapPos, getBalancesPos := analyzePrimaryReadWrap(t, src, "executeCreateTransaction")
-
-		if wrapPos == -1 {
-			t.Fatal("no `readrouting.WithPrimaryRead(ctx)` wrap found in executeCreateTransaction; the create flow does not mark the primary-read intent")
-		}
-
-		if getBalancesPos == -1 {
-			t.Fatal("no GetBalances call found in executeCreateTransaction; the read call site moved")
-		}
-
-		if wrapPos >= getBalancesPos {
-			t.Errorf("the WithPrimaryRead wrap (stmt %d) must precede the GetBalances read (stmt %d) so both pre-write balance reads observe the mark", wrapPos, getBalancesPos)
-		}
-	})
-}
-
-// primaryReadCapture records what the balance DB seam observed.
-type primaryReadCapture struct {
-	seen        bool
-	primaryRead bool
-}
-
-// newPrimaryReadCapturingUseCase builds a real query.UseCase whose Redis cache
-// always misses (forcing GetBalances to fall through to the DB seam) and whose
-// balance DB repository records the primary-read intent of the ctx it receives.
-func newPrimaryReadCapturingUseCase(ctrl *gomock.Controller) (*primaryReadCapture, *query.UseCase) {
-	captured := &primaryReadCapture{}
-
-	mockRedis := redis.NewMockRedisRepository(ctrl)
-	mockBalance := balance.NewMockRepository(ctrl)
-
-	// Cache miss on every alias: empty value -> GetBalances falls through to DB.
-	mockRedis.EXPECT().
-		Get(gomock.Any(), gomock.Any()).
-		Return("", nil).
-		AnyTimes()
-
-	mockBalance.EXPECT().
-		ListByAliasesWithKeys(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).
-		DoAndReturn(func(ctx context.Context, _, _ uuid.UUID, aliases []string) ([]*mmodel.Balance, error) {
-			captured.seen = true
-			captured.primaryRead = readrouting.IsPrimaryRead(ctx)
-
-			out := make([]*mmodel.Balance, 0, len(aliases))
-			for _, a := range aliases {
-				alias, _, _ := strings.Cut(a, "#")
-				out = append(out, companionOverdraftBalance(alias))
-			}
-
-			return out, nil
-		}).
-		AnyTimes()
-
-	uc := &query.UseCase{
-		BalanceRepo:          mockBalance,
-		TransactionRedisRepo: mockRedis,
+	if positions.wrap == -1 {
+		t.Fatal("no dedicated `readCtx := readrouting.WithPrimaryRead(ctx)` wrap found in commitOrCancelTransaction; the commit/cancel flow must mark the primary-read intent on a dedicated ctx var, not reassign ctx")
 	}
 
-	return captured, uc
-}
-
-// analyzePrimaryReadWrap returns the top-level statement indices, within the
-// named function body, of the `readrouting.WithPrimaryRead(...)` wrap and the
-// first `GetBalances(...)` call. Either is -1 when absent. Statement indices are
-// sufficient because both live at the top level of the strictly-sequential
-// create flow.
-func analyzePrimaryReadWrap(t *testing.T, src, funcName string) (wrapPos, getBalancesPos int) {
-	t.Helper()
-
-	fset := token.NewFileSet()
-
-	file, err := parser.ParseFile(fset, "src.go", src, 0)
-	if err != nil {
-		t.Fatalf("parse source: %v", err)
+	if positions.getBalances == -1 {
+		t.Fatal("no GetBalances call found in commitOrCancelTransaction; the read call site moved")
 	}
 
-	var fn *ast.FuncDecl
-
-	for _, decl := range file.Decls {
-		if d, ok := decl.(*ast.FuncDecl); ok && d.Name.Name == funcName {
-			fn = d
-			break
-		}
+	if positions.getLedgerSettings == -1 {
+		t.Fatal("no GetParsedLedgerSettings call found in commitOrCancelTransaction; the validation-only read moved")
 	}
 
-	if fn == nil || fn.Body == nil {
-		t.Fatalf("function %q not found or has no body", funcName)
+	if positions.wrap >= positions.getBalances {
+		t.Errorf("the readCtx wrap (stmt %d) must precede the GetBalances read (stmt %d) so both pre-write balance reads observe the mark", positions.wrap, positions.getBalances)
 	}
 
-	wrapPos, getBalancesPos = -1, -1
-
-	for i, stmt := range fn.Body.List {
-		if wrapPos == -1 && stmtCallsSelector(stmt, "readrouting", "WithPrimaryRead") {
-			wrapPos = i
-		}
-
-		if getBalancesPos == -1 && stmtCallsMethod(stmt, "GetBalances") {
-			getBalancesPos = i
-		}
+	if positions.wrap <= positions.getLedgerSettings {
+		t.Errorf("the readCtx wrap (stmt %d) must FOLLOW the validation-only GetParsedLedgerSettings read (stmt %d) so validation-only reads are NOT marked", positions.wrap, positions.getLedgerSettings)
 	}
 
-	return wrapPos, getBalancesPos
-}
+	// Arg identity: the marker must be scoped to the pre-write balance reads.
+	if !positions.getBalancesTakesReadCtx {
+		t.Error("the GetBalances read must receive the dedicated readCtx (the primary-read marker); passing the unmarked ctx would stop routing the balance seed to primary")
+	}
 
-// stmtCallsSelector reports whether the statement contains a call of the form
-// pkg.Fn(...), matching both the package identifier and the function name.
-func stmtCallsSelector(stmt ast.Stmt, pkg, fn string) bool {
-	found := false
+	if !positions.enrichTakesReadCtx {
+		t.Error("the cancel EnrichOverdraftOperations read must receive the dedicated readCtx so its internal GetBalances is routed to primary")
+	}
 
-	ast.Inspect(stmt, func(n ast.Node) bool {
-		call, ok := n.(*ast.CallExpr)
-		if !ok {
-			return true
-		}
-
-		if sel, ok := call.Fun.(*ast.SelectorExpr); ok && sel.Sel.Name == fn {
-			if id, ok := sel.X.(*ast.Ident); ok && id.Name == pkg {
-				found = true
-			}
-		}
-
-		return true
-	})
-
-	return found
-}
-
-// TestCommitCancel_MarksPrimaryReadIntent locks the decision that the
-// commit/cancel state-transition flow marks its pre-write balance read with the
-// primary-read intent. The read at the GetBalances call site feeds
-// buildBalanceOperations and (on cancel) enrichOverdraftOperations, whose result
-// seeds the authoritative balance via the NX-seed — the stale-read money-
-// corruption scenario the seam guards against. It asserts:
-//
-//  1. Mechanism: the intent marker survives the exact call sequence the handler
-//     uses — GetBalances and the cancel overdraft-enrichment read (which reuses
-//     the SAME ctx) both forward a marked ctx down to the balance DB seam target.
-//     A negative baseline proves an unmarked ctx stays unmarked.
-//  2. Placement: a structural guard over the live source of
-//     commitOrCancelTransaction proves the dedicated-var wrap
-//     `readCtx := readrouting.WithPrimaryRead(ctx)` exists AND sits AFTER the
-//     validation-only reads (GetParsedLedgerSettings) and BEFORE the GetBalances
-//     read; that GetBalances and the cancel overdraft read receive readCtx while
-//     ValidateAccountingRules keeps the unmarked ctx, so only the pre-write
-//     balance reads are routed to primary.
-func TestCommitCancel_MarksPrimaryReadIntent(t *testing.T) {
-	t.Run("direct_balance_read_forwards_marked_ctx_to_db_seam", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
-
-		organizationID := uuid.New()
-		ledgerID := uuid.New()
-		aliases := []string{"@alice#default"}
-
-		ctx := readrouting.WithPrimaryRead(context.Background())
-
-		_, err := uc.GetBalances(ctx, organizationID, ledgerID, aliases)
-		require.NoError(t, err)
-
-		require.True(t, captured.seen, "balance DB repository was never reached; the cache-miss path must fall through to the seam target")
-		assert.True(t, captured.primaryRead, "the ctx reaching the balance DB seam must carry the primary-read intent")
-	})
-
-	t.Run("cancel_overdraft_enrichment_read_forwards_marked_ctx_to_db_seam", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
-
-		handler := &TransactionHandler{Query: uc}
-
-		orgID := uuid.New()
-		ledgerID := uuid.New()
-
-		source := overdraftEnabledBalance(t, "@alice", decimal.NewFromInt(50), "100")
-
-		primary := mmodel.BalanceOperation{
-			Balance: source,
-			Alias:   "0#@alice#default",
-			Amount: mtransaction.Amount{
-				Asset:           "BRL",
-				Value:           decimal.NewFromInt(100),
-				Operation:       libConstants.DEBIT,
-				TransactionType: libConstants.CREATED,
-				Direction:       constant.DirectionCredit,
-			},
-			InternalKey: utils.BalanceInternalKey(orgID, ledgerID, "@alice#default"),
-		}
-
-		validate := &mtransaction.Responses{
-			From:    map[string]mtransaction.Amount{"0#@alice#default": primary.Amount},
-			Sources: []string{"@alice#default"},
-			Aliases: []string{"@alice#default"},
-		}
-
-		// The dedicated readCtx at the GetBalances call site also covers the cancel
-		// overdraft-enrichment read, which receives the same readCtx and uses
-		// handler.Query.GetBalances as its loader exactly as at the handler site.
-		ctx := readrouting.WithPrimaryRead(context.Background())
-
-		_, _, err := enrichOverdraftOperations(ctx, orgID, ledgerID,
-			[]mmodel.BalanceOperation{primary}, validate, handler.Query.GetBalances)
-		require.NoError(t, err)
-
-		require.True(t, captured.seen, "cancel overdraft companion read never reached the balance DB seam")
-		assert.True(t, captured.primaryRead, "the ctx reaching the balance DB seam via the cancel overdraft read must carry the primary-read intent")
-	})
-
-	t.Run("unmarked_baseline_ctx_is_not_marked", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		defer ctrl.Finish()
-
-		captured, uc := newPrimaryReadCapturingUseCase(ctrl)
-
-		_, err := uc.GetBalances(context.Background(), uuid.New(), uuid.New(), []string{"@alice#default"})
-		require.NoError(t, err)
-
-		require.True(t, captured.seen)
-		assert.False(t, captured.primaryRead, "a flow not going through commit/cancel must NOT carry the primary-read intent")
-	})
-
-	t.Run("dedicated_readctx_wrap_is_scoped_to_the_pre_write_balance_reads", func(t *testing.T) {
-		src := readStateHandlerSource(t)
-
-		positions := analyzeCommitCancelWrap(t, src, "commitOrCancelTransaction")
-
-		if positions.wrap == -1 {
-			t.Fatal("no dedicated `readCtx := readrouting.WithPrimaryRead(ctx)` wrap found in commitOrCancelTransaction; the commit/cancel flow must mark the primary-read intent on a dedicated ctx var, not reassign ctx")
-		}
-
-		if positions.getBalances == -1 {
-			t.Fatal("no GetBalances call found in commitOrCancelTransaction; the read call site moved")
-		}
-
-		if positions.getLedgerSettings == -1 {
-			t.Fatal("no GetParsedLedgerSettings call found in commitOrCancelTransaction; the validation-only read moved")
-		}
-
-		if positions.wrap >= positions.getBalances {
-			t.Errorf("the readCtx wrap (stmt %d) must precede the GetBalances read (stmt %d) so both pre-write balance reads observe the mark", positions.wrap, positions.getBalances)
-		}
-
-		if positions.wrap <= positions.getLedgerSettings {
-			t.Errorf("the readCtx wrap (stmt %d) must FOLLOW the validation-only GetParsedLedgerSettings read (stmt %d) so validation-only reads are NOT marked", positions.wrap, positions.getLedgerSettings)
-		}
-
-		// Arg identity: the marker must be scoped to the pre-write balance reads.
-		if !positions.getBalancesTakesReadCtx {
-			t.Error("the GetBalances read must receive the dedicated readCtx (the primary-read marker); passing the unmarked ctx would stop routing the balance seed to primary")
-		}
-
-		if !positions.enrichTakesReadCtx {
-			t.Error("the cancel enrichOverdraftOperations read must receive the dedicated readCtx so its internal GetBalances is routed to primary")
-		}
-
-		if positions.validateTakesReadCtx {
-			t.Error("ValidateAccountingRules must NOT receive the dedicated readCtx: it is validation-only (route-cache) and deliberately keeps the unmarked ctx")
-		}
-	})
+	if positions.validateTakesReadCtx {
+		t.Error("ValidateAccountingRules must NOT receive the dedicated readCtx: it is validation-only (route-cache) and deliberately keeps the unmarked ctx")
+	}
 }
 
 // commitCancelWrapPositions holds the top-level statement indices of the marker
@@ -410,7 +77,7 @@ type commitCancelWrapPositions struct {
 // analyzeCommitCancelWrap returns, within the named function body, the top-level
 // statement indices of the dedicated `readCtx := readrouting.WithPrimaryRead(...)`
 // wrap, the first GetBalances call, and the validation-only GetParsedLedgerSettings
-// call (each -1 when absent), and whether GetBalances, enrichOverdraftOperations,
+// call (each -1 when absent), and whether GetBalances, EnrichOverdraftOperations,
 // and ValidateAccountingRules receive `readCtx` as their context argument.
 // Statement indices are sufficient because the reads live at the top level of the
 // sequential commit/cancel flow.
@@ -456,7 +123,7 @@ func analyzeCommitCancelWrap(t *testing.T, src, funcName string) commitCancelWra
 			positions.getBalancesTakesReadCtx = true
 		}
 
-		if callFirstArgIsIdent(stmt, "enrichOverdraftOperations", "readCtx") {
+		if callFirstArgIsIdent(stmt, "EnrichOverdraftOperations", "readCtx") {
 			positions.enrichTakesReadCtx = true
 		}
 
@@ -500,7 +167,7 @@ func stmtDefinesReadCtxFromPrimaryRead(stmt ast.Stmt) bool {
 // callFirstArgIsIdent reports whether stmt contains a call to callee whose first
 // argument is exactly the identifier argName. callee matches either a method call
 // (`x.callee(...)`) or a plain function call (`callee(...)`), so both
-// `handler.Query.GetBalances(readCtx, ...)` and `enrichOverdraftOperations(readCtx, ...)`
+// `handler.Query.GetBalances(readCtx, ...)` and `command.EnrichOverdraftOperations(readCtx, ...)`
 // are recognized — used to assert which reads receive the dedicated readCtx versus
 // the unmarked ctx.
 func callFirstArgIsIdent(stmt ast.Stmt, callee, argName string) bool {
@@ -556,4 +223,25 @@ func readStateHandlerSource(t *testing.T) string {
 	}
 
 	return src
+}
+
+// stmtCallsMethod reports whether the statement contains a selector call whose
+// method name matches (e.g. handler.Query.GetBalances(...)).
+func stmtCallsMethod(stmt ast.Stmt, method string) bool {
+	found := false
+
+	ast.Inspect(stmt, func(n ast.Node) bool {
+		call, ok := n.(*ast.CallExpr)
+		if !ok {
+			return true
+		}
+
+		if sel, ok := call.Fun.(*ast.SelectorExpr); ok && sel.Sel.Name == method {
+			found = true
+		}
+
+		return true
+	})
+
+	return found
 }
