+++ components/ledger/internal/adapters/http/in/transaction_revert_no_refund_test.go
@@ -11,67 +11,15 @@ import (
 	"os"
 	"testing"
 
-	"github.com/google/uuid"
-	"github.com/shopspring/decimal"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/LerianStudio/midaz/v4/components/ledger/internal/adapters/tracer"
-	"github.com/LerianStudio/midaz/v4/pkg/mmodel"
 )
 
-// TestRevertNoReservationRefund is the permanent behavioral lock for Q9
-// (no-refund on revert): reverting a transaction must NEVER Release or Confirm
-// the ORIGINAL transaction's reservation. Limits measure GROSS activity, so a
-// revert is itself a new chargeable transaction that reserves on its own via
-// the create anchor (F3-T13); the original reservation is left exactly as the
-// original transaction left it (confirmed on the original commit).
-//
-// The lock has two halves:
-//
-//  1. A structural guard (TestRevertNoReservationRefund_StructuralGuard) over
-//     transaction_state_handlers.go asserting the revert entry points contain
-//     no Release/Confirm call against any reservation. A future change that
-//     added a refund would have to add such a call there and would trip it.
-//  2. This behavioral test: the reverse transaction reserves on its own and
-//     the original reservation id is never touched.
-func TestRevertNoReservationRefund(t *testing.T) {
-	ctx, sp, logger := anchorDeps()
-
-	// The "original" transaction's reservation — the one a buggy refund would
-	// release/confirm. We assert it is never referenced.
-	originalReservationID := uuid.New()
-
-	// The revert path delegates to executeCreateTransaction, whose reserve
-	// anchor issues a NEW reserve for the reverse transaction. Model that the
-	// reverse transaction reserves on its own and capture which ids the ledger
-	// later confirms/releases.
-	reverseReservationID := uuid.New()
-	reserver := &stubReserver{result: &tracer.ReserveResult{ReservationIDs: []uuid.UUID{reverseReservationID}}}
-	handler := &TransactionHandler{TracerReserver: reserver}
-
-	// Reserve for the reverse transaction (what the revert's executeCreateTransaction does).
-	out := handler.reserveTransaction(ctx, sp, logger,
-		mmodel.TracerSettings{Mode: mmodel.TracerModeEnforce, FailPosture: mmodel.TracerFailPostureOpen},
-		uuid.New(), decimal.NewFromInt(1000), "BRL", fixedReserveAccountID, fixedReserveTimestamp, reservationTTLDefault, routeV2, false)
-	require.Equal(t, reservationProceed, out.Kind)
-
-	// On a successful reverse-transaction commit the ledger confirms the
-	// REVERSE reservation — never the original.
-	handler.confirmReservations(ctx, sp, logger, out.Handle)
-
-	require.Equal(t, []uuid.UUID{reverseReservationID}, reserver.confirmedIDs,
-		"a revert confirms its own reverse-transaction reservation")
-	assert.NotContains(t, reserver.confirmedIDs, originalReservationID,
-		"a revert must NEVER confirm the original transaction's reservation")
-	assert.NotContains(t, reserver.releasedIDs, originalReservationID,
-		"a revert must NEVER release the original transaction's reservation (Q9 no-refund)")
-}
-
-// TestRevertNoReservationRefund_StructuralGuard asserts the revert entry points
-// in transaction_state_handlers.go invoke neither Release nor Confirm. The
-// confirm/release transport lives only in the create seam
-// (transaction_create.go); the state handlers must stay refund-free.
+// TestRevertNoReservationRefund_StructuralGuard asserts the revert core in
+// transaction_state_handlers.go invokes neither Release nor Confirm. The
+// confirm/release transport lives only in the create seam; the revert core must stay
+// refund-free, because limits measure GROSS activity: a revert reserves on its own and
+// never refunds the original transaction's reservation (Q9 no-refund). The behavioral
+// half of the lock lives beside the reserve anchor, in the command package.
 func TestRevertNoReservationRefund_StructuralGuard(t *testing.T) {
 	src, err := os.ReadFile("transaction_state_handlers.go")
 	require.NoError(t, err, "read transaction_state_handlers.go")
@@ -81,15 +29,17 @@ func TestRevertNoReservationRefund_StructuralGuard(t *testing.T) {
 	file, err := parser.ParseFile(fset, "transaction_state_handlers.go", src, 0)
 	require.NoError(t, err, "parse transaction_state_handlers.go")
 
-	// Walk RevertTransaction and createRevertTransaction (the revert entry
-	// points) and assert no Release/Confirm method call appears in either.
-	revertFuncs := map[string]bool{"RevertTransaction": true, "createRevertTransaction": true}
+	// Walk the revert core and assert no Release/Confirm method call appears in it.
+	revertFuncs := map[string]bool{"revertTransaction": true}
+	seen := false
 
 	for _, decl := range file.Decls {
 		fn, ok := decl.(*ast.FuncDecl)
 		if !ok || !revertFuncs[fn.Name.Name] {
 			continue
 		}
+
+		seen = true
 
 		ast.Inspect(fn, func(n ast.Node) bool {
 			call, ok := n.(*ast.CallExpr)
@@ -110,4 +60,6 @@ func TestRevertNoReservationRefund_StructuralGuard(t *testing.T) {
 			return true
 		})
 	}
+
+	require.True(t, seen, "revertTransaction not found in transaction_state_handlers.go — the guard is pointed at a renamed core")
 }
