+++ components/ledger/internal/services/command/build_transaction_operations_snapshot_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the Elastic License 2.0
 // that can be found in the LICENSE file.
 
-package in
+package command
 
 import (
 	"testing"
