+++ components/ledger/internal/services/command/transaction_helper_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the Elastic License 2.0
 // that can be found in the LICENSE file.
 
-package in
+package command
 
 import (
 	"context"
@@ -49,7 +49,7 @@ func TestGetAliasWithoutKey(t *testing.T) {
 		"third#asset#extra",
 	}
 
-	result := getAliasWithoutKey(input)
+	result := GetAliasWithoutKey(input)
 
 	assert.Equal(t, []string{"origin", "destination", "third"}, result)
 }
@@ -108,7 +108,7 @@ func TestBuildBalanceOperations(t *testing.T) {
 			},
 		}
 
-		ops := buildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
+		ops := BuildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
 		require.Len(t, ops, 1)
 		assert.Equal(t, "0#alias1#default", ops[0].Alias)
 		assert.Equal(t, constant.DEBIT, ops[0].Amount.Operation)
@@ -150,7 +150,7 @@ func TestBuildBalanceOperations(t *testing.T) {
 			},
 		}
 
-		ops := buildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
+		ops := BuildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
 		require.Len(t, ops, 2)
 		assert.Equal(t, ops[0].Alias, ops[1].Alias, "both operations should reference the same alias")
 	})
@@ -189,7 +189,7 @@ func TestBuildBalanceOperations(t *testing.T) {
 			},
 		}
 
-		ops := buildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
+		ops := BuildBalanceOperations(context.Background(), organizationID, ledgerID, validate, balances)
 		require.Len(t, ops, 2)
 
 		// Verify sorted by internal key
